BUG: date comparison fails when series is all pd.NaT values (#61200)

+++ pandas/core/indexes/accessors.py
@@ -108,7 +108,9 @@ class Properties(PandasDelegate, PandasObject, NoNewAttributesMixin):
         else:
             index = self._parent.index
         # return the result as a Series
-        return Series(result, index=index, name=self.name).__finalize__(self._parent)
+        return Series(
+            result, index=index, name=self.name, dtype=result.dtype
+        ).__finalize__(self._parent)
 
     def _delegate_property_set(self, name: str, value, *args, **kwargs) -> NoReturn:
         raise ValueError(
